+++ Merchant/V1/index.ts
@@ -1,0 +1,5 @@
+import { Key } from "./Key"
+
+export {
+	Key,
+}
